[Fix] 카테고리 수정

+++ app/api/category/service.py
@@ -1,5 +1,5 @@
 from sqlalchemy.ext.asyncio import AsyncSession
-from app.api.category.schema import CreateCategoryRequest, CategoryResponse, UpdateCategoryRequest
+from app.api.category.schema import CreateCategoryRequest, CategoryResponse, UpdateCategoryRequest, CategoryResponseList
 from app.api.category import repository
 from app.exceptions.custom_exception import CustomException
 
@@ -39,6 +39,16 @@ async def delete_category(session: AsyncSession, category_id: int, user_id: int)
             
             if not db_category:
                 raise CustomException("존재하지 않는 카테고리")
+
+            if db_category.name == "기타":
+                raise CustomException("기타 카테고리는 삭제할 수 없습니다")
+            
+            scrap_count = await repository.count_scraps_in_category(
+                session, category_id=category_id
+            )
+
+            if scrap_count > 0:
+                raise CustomException("카테고리에 스크랩이 존재하여 삭제할 수 없습니다")
             
             await repository.delete_category(session, db_category)
             return db_category
@@ -51,3 +61,16 @@ async def delete_category(session: AsyncSession, category_id: int, user_id: int)
         logger.error("error : {}", sys.exc_info())
 
         raise CustomException("카테고리 삭제 실패")
+    
+async def get_category_by_user_id(session: AsyncSession, user_id: int) -> CategoryResponseList | None:
+    try:
+        async with session as session:
+            category = await repository.get_categories_by_user_id(session, user_id)
+            
+            if not category:
+                return None
+            
+            return category
+    except Exception as e:
+        logger.error("error : {}", sys.exc_info())
+        raise CustomException("카테고리 조회 실패")
